Speed up checking for move legality
Before, move legality checking was done by making each move,
and then checking if the moving side's king was left in check
by calling sqIsAttacked. This is inefficent, as only four
types of moves need to be checked:

    (1) Check evasion moves
    (2) King moves
    (3) En passant moves
    (4) Pinned piece moves

So, the getPinnedPieces method was created to calculate pinned
pieces, and logic to detect the other conditions was added to
makeMove, so now sqIsAttacked is more selectively called.

This increased PERFT speed by roughly 10-12%.