Speed-up movegen by limiting pinned piece computation
Rathering than computing a new pinned pieced bitboard for each call
to Position.makeMove, compute the bitboard once for a parent position,
and pass the information along to makeMove calls on the parent's children.
This ensures that time is spent calculating pinned pieces for the new child
position only if it's legal and will actually be visited.