Improve algorithm for getPieceType and getPieceColor
Before, due to the algorithm used in getPieceType and getPieceColor,
the NO_TYPE and NO_COLOR constants had to be set to zero, as those
were the values the two methods would return if there was no piece
on a square. This meant that each indexing constant used for pieces
and colors has to be shifted upwards by 1. So PAWN = 1, KNIGHT = 2,
etc.

This meant memory had to be wasted when creating arrays that would
be indexed by these piece constants, since they would need to be
made one size bigger to make sure the piece constants could index
them correctly.

After improving the algorithm in getPieceType and getPieceColor,
both methods are now able to return values different than 0 when
a piece is not found on a square, meaning piece constants can now
start at 0, and the arrays which they index can be reduced in size,
saving time and memory.